Extract formatWithDecimals()
