+++ src/commonMain/kotlin/nl/jacobras/humanreadable/DecimalsUtil.kt
@@ -1,0 +1,17 @@
+package nl.jacobras.humanreadable
+
+import kotlin.math.pow
+import kotlin.math.roundToLong
+
+internal fun Float.formatWithDecimals(decimals: Int): String {
+    val multiplier = 10.0.pow(decimals)
+    val numberAsString = (this * multiplier).roundToLong().toString()
+    val decimalIndex = numberAsString.length - decimals - 1
+    val mainRes = numberAsString.substring(0..decimalIndex)
+    val fractionRes = numberAsString.substring(decimalIndex + 1)
+    return if (fractionRes.isEmpty()) {
+        mainRes
+    } else {
+        "$mainRes.$fractionRes"
+    }
+}
