+++ src/commonMain/kotlin/nl/jacobras/humanreadable/HumanReadableFileSize.kt
@@ -1,8 +1,6 @@
 package nl.jacobras.humanreadable
 
 import Res
-import kotlin.math.pow
-import kotlin.math.roundToLong
 
 /**
  * Returns the given [bytes] size in human-readable format.
@@ -24,18 +22,5 @@ internal fun formatFileSize(bytes: Long, decimals: Int): String {
         else -> {
             "${(bytes / 1.09951163E12f).formatWithDecimals(decimals)} ${Res.string.terabyte_symbol}"
         }
-    }
-}
-
-private fun Float.formatWithDecimals(decimals: Int): String {
-    val multiplier = 10.0.pow(decimals)
-    val numberAsString = (this * multiplier).roundToLong().toString()
-    val decimalIndex = numberAsString.length - decimals - 1
-    val mainRes = numberAsString.substring(0..decimalIndex)
-    val fractionRes = numberAsString.substring(decimalIndex + 1)
-    return if (fractionRes.isEmpty()) {
-        mainRes
-    } else {
-        "$mainRes.$fractionRes"
     }
 }
